Rename Controller.java and implement its changes
Renamed Controller.java to XBoxController.java so that way no names
clashed, and it would still be fairly descriptive as to what the class
actually is. The XBoxController was also implemented into the
RobotContainer class. The single method for getting a trigger was
split into 3 different methods. One for getting the buttons, one for
getting the triggers, and one for getting the axes. This was so that
way you could actually use some of the specifics on each abstraction
without getting a compiler error.