Finish creation of SwerveModule class
Change constructor to automatically pick the motors and controller
based only on an id passed in so instead of passing in 3 motors, only
pass in 1 integer. Added methods to get offset of the module, heading,
and set the state that the module should be in. Also implemented an
optimize function that automatically deals with continuous headings.