Improve and bug fix auto
Fixed an error that caused the robot to not follow the path correctly
and changed the paths a bit to allow for easier tuning in the future.
Also made a plan to allow for easier creation of auto commands, and
those will be implemented shortly with general refactoring around it as
well.