Improve path following and change hood setpoints
Changed odometry in a way that the robot is where it thinks it is, and
improved path following in that way. The gear ratio on the hood was
also changed, so setpoints needed to be adjusted for that change.
Setpoints will need retuned, but, for now, the current points work.
Limelight calibration not changed, but will be retuned later.