Move limiters to DriveCommand, fix auto issues, and clean up swerveModule
Moved the SlewRateLimiters to the DriveCommand so
it wouldn't interfere with Auto. Also fixed an issue
where if the heading of the initial pose in the path
wasn't the same as the rotation, the robot would
move in the wrong direction. Cleaned up the
SwerveModule.optimize method, and removed the
SlewRateLimiter from it.